core/levels/defender/audit/resources/rmUser/requirements.txt
google-auth
google-cloud-logging
